+++ src/main/kotlin/com/example/autobank/data/receipt/CompleteReceipt.kt
@@ -1,10 +1,5 @@
 package com.example.autobank.data.receipt
 
-
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.Id
-import jakarta.persistence.Table
 import java.math.BigDecimal
 import java.time.LocalDateTime
