+++ src/main/kotlin/com/example/autobank/data/receipt/ReceiptListResponseBody.kt
@@ -1,5 +1,7 @@
 package com.example.autobank.data.receipt
 
+import com.example.autobank.data.models.ReceiptInfo
+
 class ReceiptListResponseBody(
     val receipts: Array<ReceiptInfo>,
     val total: Long,
