+++ src/main/kotlin/com/example/autobank/repository/CommitteeRepository.kt
@@ -1,7 +1,6 @@
 package com.example.autobank.repository;
 
-import com.example.autobank.data.Committee
-import com.example.autobank.data.Economicrequest
+import com.example.autobank.data.models.Committee
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
