+++ src/main/kotlin/com/example/autobank/repository/EconomicRequestReviewRepository.kt
@@ -1,6 +1,6 @@
 package com.example.autobank.repository
 
-import com.example.autobank.data.EconomicRequestReview
+import com.example.autobank.data.models.EconomicRequestReview
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Modifying
 import org.springframework.data.jpa.repository.Query
