+++ src/main/kotlin/com/example/autobank/repository/EconomicrequestRepository.kt
@@ -1,6 +1,6 @@
 package com.example.autobank.repository;
 
-import com.example.autobank.data.Economicrequest
+import com.example.autobank.data.models.Economicrequest
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
