+++ src/main/kotlin/com/example/autobank/repository/receipt/AttachmentRepository.kt
@@ -1,6 +1,6 @@
 package com.example.autobank.repository.receipt
 
-import com.example.autobank.data.receipt.Attachment
+import com.example.autobank.data.models.Attachment
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
