+++ src/main/kotlin/com/example/autobank/repository/receipt/CardRepository.kt
@@ -1,6 +1,6 @@
 package com.example.autobank.repository.receipt
 
-import com.example.autobank.data.receipt.Card
+import com.example.autobank.data.models.Card
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
