+++ src/main/kotlin/com/example/autobank/repository/receipt/ReceiptInfoViewRepository.kt
@@ -1,6 +1,6 @@
 package com.example.autobank.repository.receipt
 
-import com.example.autobank.data.receipt.ReceiptInfo
+import com.example.autobank.data.models.ReceiptInfo
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.data.jpa.repository.JpaRepository
