+++ src/main/kotlin/com/example/autobank/repository/receipt/ReceiptRepository.kt
@@ -1,6 +1,6 @@
 package com.example.autobank.repository.receipt
 
-import com.example.autobank.data.receipt.Receipt
+import com.example.autobank.data.models.Receipt
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
