+++ src/main/kotlin/com/example/autobank/repository/receipt/ReceiptReviewRepository.kt
@@ -1,6 +1,6 @@
 package com.example.autobank.repository.receipt
 
-import com.example.autobank.data.receipt.ReceiptReview
+import com.example.autobank.data.models.ReceiptReview
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 import org.springframework.data.jpa.repository.Modifying
