+++ src/main/kotlin/com/example/autobank/service/AttachmentService.kt
@@ -1,6 +1,6 @@
 package com.example.autobank.service
 
-import com.example.autobank.data.receipt.Attachment
+import com.example.autobank.data.models.Attachment
 import com.example.autobank.repository.receipt.AttachmentRepository
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
